// frontend/src/models/order.model.ts
export interface Order {
    id: number;
    createdAt: string;
    address: string;
    receiver: string;
    contact: string;
    bookTitle: string;
    totalQuantity: number;
    totalPrice: number;
}

export interface OrderSheet {
    items: number[];
    totalQuantity: number;
    totalPrice: number;
    firstBookTitle: string;
    delivery: Delivery;
}

export interface Delivery {
    address: string;
    receiver: string;
    contact: string;
}

export interface OrderDetailItem {
    bookId: number;
    title: string;
    author: string;
    price: number;
    quantity: number;
}

export interface OrderListItem extends Order {
    detail?: OrderDetailItem[];
}